Move SDMM/legacy mount functions into different file to fix binary bloat
